refactor(ui): extract messages API into separate module
Move message-related functions from sessions.ts to new messages.ts file
for better separation of concerns. Re-export from sessions.ts for
backward compatibility.
