+++ apps/ui/src/lib/api/index.ts
@@ -3,5 +3,6 @@ export * from "./types";
 export * from "./client";
 export * from "./agents";
 export * from "./sessions";
+export * from "./messages";
 export * from "./llm-providers";
 export * from "./auth";
